Electro-Pneumatic Simulator (EPSim) At this stage, users should be able to create electro-pneumatic circuits with a ladder diagram similar to a Programmable Logic Controller. Wholly pneumatic circuits are not supported.
Lab 2 and Lab 3 csv files good indicators of the current capabilities. Generally 'run with step' is used together with the START and/or STOP push buttons.
Lab 3 is a Cascade Method implementation of a2 b2 a1 c2 c1 d2 b1 d1 which is the sequence A+ B+ A- C+ C- D+ B- D-.